Add example submission script
In an effort to clarify the process for conformance submission,
I provided an example script which just centralizes the steps
discussed in the document and also puts into code some of the
steps which were just talked about in sentence form.

I also updated a few of the Sonobuoy statements to reflect new
options and behavior.