Update instructions for certified conformance runs via Sonobuoy
As of the upcoming v1.16.0 Kubernetes release, some disruptive tests
are being put into the conformance suite. This is necessary to show
full conformance, but for end-users running tests on their clusters
it is a potentially harmful situation.

To mitigate this, Sonobuoy has made the default `sonobuoy run` command
run the set of non-disruptive-conformance tests. This is suitable for
any running cluster, but fails to run 100% of the conformance tests
and, as a result, is invalid for conformance certification.

Sonobuoy v0.15.4 added a new option to specify that you want to run
the certification suite, which would include the disruptive tests:
`sonobuoy run --mode=certified-conformance`

Signed-off-by: John Schnake <jschnake@vmware.com>