Make a note about recent Sonobuoy issue (#1570)
There was a recent bug reported to Sonobuoy that caused a handful
of requests for certification to fail due to a missing test. Sonoboy
has already patchd the issue but I wanted to add a note here to
ensure users following these instructions have a chance to avoid it first.

Signed-off-by: John Schnake <jschnake@vmware.com>
Signed-off-by: Avinash Kumar Singh <avinash@xenonstack.com>